+++ ext/LieGroupsRecursiveArrayToolsExt/LieGroupsRecursiveArrayToolsExt.jl
@@ -4,6 +4,7 @@ using LieGroups
 using RecursiveArrayTools: ArrayPartition
 using LinearAlgebra
 using ManifoldsBase
+using ManifoldsBase: base_manifold
 
 include("special_euclidean_group_RAT_ext.jl")